Moved provider-specific shared utils to its own file. The intention is that frameworkProvider.go will absorb all this when sdkProvider is removed
