--- provider/debug_logger.go
@@ -1,21 +1,0 @@
-package provider
-
-import (
-	"fmt"
-	"log"
-	"strings"
-)
-
-type DebugLogger struct{}
-
-func (d *DebugLogger) Printf(format string, v ...interface{}) {
-	log.Printf("[DEBUG] [rediscloud-go-api] "+format, v...)
-}
-
-func (d *DebugLogger) Println(v ...interface{}) {
-	var items []string
-	for _, i := range v {
-		items = append(items, fmt.Sprintf("%s", i))
-	}
-	log.Printf("[DEBUG] [rediscloud-go-api] %s", strings.Join(items, " "))
-}
