+++ provider/utils.go
@@ -2,21 +2,12 @@ package provider
 
 import (
 	"fmt"
-	rediscloudApi "github.com/RedisLabs/rediscloud-go-api"
 	"github.com/RedisLabs/rediscloud-go-api/redis"
 	"github.com/hashicorp/go-cty/cty"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"sync"
 	"time"
 )
-
-// Lock that must be acquired when modifying something related to a subscription as only one _thing_ can modify a subscription and all sub-resources at any time
-var subscriptionMutex = newPerIdLock()
-
-type ApiClient struct {
-	client *rediscloudApi.Client
-}
 
 func setToStringSlice(set *schema.Set) []*string {
 	var ret []*string
@@ -32,38 +23,6 @@ func interfaceToStringSlice(list []interface{}) []*string {
 		ret = append(ret, redis.String(i.(string)))
 	}
 	return ret
-}
-
-type perIdLock struct {
-	lock  sync.Mutex
-	store map[int]*sync.Mutex
-}
-
-func newPerIdLock() *perIdLock {
-	return &perIdLock{
-		store: map[int]*sync.Mutex{},
-	}
-}
-
-func (m *perIdLock) Lock(id int) {
-	m.get(id).Lock()
-}
-
-func (m *perIdLock) Unlock(id int) {
-	m.get(id).Unlock()
-}
-
-func (m *perIdLock) get(id int) *sync.Mutex {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	if v, ok := m.store[id]; ok {
-		return v
-	}
-
-	mutex := &sync.Mutex{}
-	m.store[id] = mutex
-	return mutex
 }
 
 // IDs of any resources dependent on a subscription need to be divided by a slash. In this format: <sub id>/<resource id>.
